Support reading De Extra 500 2023 again
The separate list can be obtained by passing `python -m top2000 2023.5`
as the calling arguments. This should also be available as a data source
(merged with the normal 2023 list) for later years, which leads to such
tracks that will come back into the main list in 2024 as being reported
as increasing from a former position below 2000 instead of new or from
earlier years further back. Maybe something special can be done to
indicate their novelty, particularly for 2025 and beyond if they do
reappear there. For now, read support for the extra list is enough.