Normalize year field intermediate naming
Similar to how artist and title are given renamed fields, ensure there
is a name where the year field is always placed, such that we can more
easily obtain it during output later on without depending on reader or
external source data. Updated outputs to only use this field for year.