HTML viewer: Initial work on progression chart
Clicking a row will add an informative panel below it, currently
displaying just a chart with positions in earlier years with a curved
line which becomes disjointed if the track has left the chart during
some years. Each year's position is marked with a point, current year is
marked with a star.