HTML viewer: Initial artist subchart symbol legend
Display which symbol at the end of the line in the progression chart
belongs to the selected artist/song. If not selected, an emoji pointing
in the direction of where the song is found is displayed.

Correct symbol range, exclude circle (used for earlier year) and wye
(weird shape and no emoji equivalent).