Parse extra years after the year itself
This sorts the extra year 2023.5 before 2023 so that it is handled later
than the full list, which avoids problems with tracks appearing twice in
the lists with slightly different artists, namely George Michael & Elton
John's "Don't Let The Sun Go Down On Me". The version that is positioned
better should be parsed first to make its best key preferred, instead of
causing collisions which should not occur.