Initial external manifest single-file version
Hacky method to allow using local data file with external resources such
that the single-file version receives updated functionality from the web
version, as long as the asset module identifier of the data file remains
the same. Currently this even has the identifier hardcoded in the custom
entry point. This needs a manifest file running on a server allowing all
origins to access it, or a JSONP wrapper which needs to be made manually
by wrapping the manifest in a "jsonptop2000" callback function.